Cleanup, additional test, javadoc
